chore: Add HasShippingAddress trait for setting shipping address in builders
